KAN-47 fix: fix css styles, password validation, rename to CreateNewPasswordForm
