--- src/features/auth/forms/CreateNewPassword/index.ts
@@ -1,1 +1,0 @@
-export * from './CreateNewPassword'
