+++ src/features/auth/forms/CreateNewPasswordForm/CreateNewPasswordForm.stories.tsx
@@ -1,0 +1,15 @@
+import type { Meta, StoryObj } from '@storybook/nextjs-vite'
+import { CreateNewPasswordForm } from './CreateNewPasswordForm'
+
+const meta = {
+   component: CreateNewPasswordForm,
+   tags: ['autodocs'],
+   title: 'Forms/CreateNewPasswordForm',
+} satisfies Meta<typeof CreateNewPasswordForm>
+
+export default meta
+type Story = StoryObj<typeof meta>
+
+export const Default: Story = {
+   args: { onSubmitAction: x => x },
+}
