+++ src/features/auth/forms/CreateNewPasswordForm/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateNewPasswordForm'
